# Kramp.API/Swagger/V1.3.0_openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Kramp API",
    "description": "Kramp API",
    "contact": {
      "name": "Página de Contato",
      "url": "https://www.exemplo.com.br"
    },
    "license": {
      "name": "Licença",
      "url": "https://www.exemplo.com.br"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Plan/Create": {
      "post": {
        "tags": [
          "Plan"
        ],
        "summary": "Cria um novo plano de treino.",
        "description": "Este endpoint permite a criação de um novo plano de treino com o nome, descrição, data de início, data de fim e treinos relacionados.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanCommand"
              },
              "examples": {
                "Exemplo de Requisição": {
                  "summary": "Um exemplo de requisição para criar um plano de treino.",
                  "value": {
                    "Name": "Plano de Treino para Iniciantes",
                    "Description": "Plano de treino para quem está começando a se exercitar.",
                    "StartDate": "2024-03-08T00:00:00Z",
                    "EndDate": "2024-04-08T00:00:00Z",
                    "Workouts": [
                      {
                        "Id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      },
                      {
                        "Id": "3fa85f64-5717-4562-b3fc-2c963f66afa7"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Plano de treino criado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanInfoViewModel"
                },
                "examples": {
                  "Exemplo de Resposta": {
                    "summary": "Um exemplo de resposta para a criação de um plano de treino.",
                    "value": {
                      "Name": "Plano de Treino para Iniciantes",
                      "Description": "Plano de treino para quem está começando a se exercitar.",
                      "StartDate": "2024-03-08T00:00:00Z",
                      "EndDate": "2024-04-08T00:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação de entrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "Valores inválidos": {
                    "summary": "Erro quando os valores fornecidos são inválidos.",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Um ou mais erros de validação ocorreram.",
                      "status": 400,
                      "detail": "Verifique os erros para mais detalhes.",
                      "instance": "/api/Plan/Create",
                      "errors": {
                        "Name": [
                          "O nome é obrigatório."
                        ],
                        "StartDate": [
                          "A data de início é obrigatória."
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "Erro interno": {
                    "summary": "Erro inesperado ao processar a requisição.",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Erro interno no servidor.",
                      "status": 500,
                      "detail": "Ocorreu um erro inesperado ao processar a sua requisição. Tente novamente mais tarde.",
                      "instance": "/api/Plan/Create"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Plan/{Id}": {
      "get": {
        "tags": [
          "Plan"
        ],
        "summary": "Retorna um plano de treino específico.",
        "description": "Este endpoint recupera os detalhes de um plano de treino específico com base no ID fornecido.",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "O UUID do plano de treino que deseja recuperar.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plano de treino encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanInfoViewModel"
                },
                "examples": {
                  "success-response": {
                    "summary": "Plano de treino encontrado com sucesso.",
                    "value": {
                      "name": "Plano de Treino A",
                      "description": "Plano para iniciantes",
                      "startDate": "2024-01-15T00:00:00Z",
                      "endDate": "2024-02-15T00:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação de entrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "invalidIdFormatError": {
                    "summary": "Erro quando o ID fornecido não está no formato UUID válido",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Formato de ID inválido.",
                      "status": 400,
                      "detail": "O formato do ID é inválido. Deve ser um UUID válido.",
                      "instance": "/api/Plan/{Id}"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plano de treino não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "planNotFoundError": {
                    "summary": "Erro quando o plano de treino com o ID fornecido não é encontrado",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Plano não encontrado.",
                      "status": 404,
                      "detail": "Plano de treino não encontrado com o ID fornecido.",
                      "instance": "/api/Plan/{Id}"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Plan/All": {
      "get": {
        "tags": [
          "Plan"
        ],
        "summary": "Retorna todos os planos de treino.",
        "description": "Este endpoint recupera uma lista de todos os planos de treino cadastrados no sistema.",
        "responses": {
          "200": {
            "description": "Sucesso - Lista de planos de treino retornada com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanInfoViewModel"
                  }
                },
                "examples": {
                  "planInfoViewModel": {
                    "summary": "Lista de planos de treino.",
                    "value": [
                      {
                        "name": "Plano de Força A",
                        "description": "Foco em hipertrofia com treinos intensos.",
                        "startDate": "2024-01-01T00:00:00Z",
                        "endDate": "2024-02-28T23:59:59Z"
                      },
                      {
                        "name": "Plano de Emagrecimento B",
                        "description": "Combinação de exercícios aeróbicos e dieta balanceada.",
                        "startDate": "2024-03-15T00:00:00Z",
                        "endDate": "2024-06-15T23:59:59Z"
                      }
                    ]
                  }
                }
              }
            }
          },
          "500": {
            "description": "Erro interno no servidor.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "internalServerError": {
                    "summary": "Erro inesperado durante o processamento da requisição.",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Erro interno do servidor.",
                      "status": 500,
                      "detail": "Ocorreu um erro inesperado ao processar a requisição. Tente novamente mais tarde.",
                      "instance": "/api/Plan/All"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Plan/Update/{Id}": {
      "put": {
        "tags": [
          "Plan"
        ],
        "summary": "Atualiza um plano de treino específico.",
        "description": "Este endpoint permite a atualização das informações de um plano de treino específico utilizando o ID fornecido.",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "O UUID do plano de treino que deseja atualizar.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePlanCommand"
              },
              "examples": {
                "valid-request": {
                  "summary": "Exemplo de requisição válida para atualizar um plano de treino.",
                  "value": {
                    "Name": "Plano de Treino Atualizado",
                    "Description": "Descrição atualizada do plano.",
                    "StartDate": "2024-01-15T00:00:00Z",
                    "EndDate": "2024-02-15T00:00:00Z",
                    "Workouts": [
                      {
                        "Id": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plano de treino atualizado com sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanInfoViewModel"
                },
                "examples": {
                  "success-response": {
                    "summary": "Exemplo de resposta com sucesso após atualizar um plano de treino.",
                    "value": {
                      "Name": "Plano de Treino Atualizado",
                      "Description": "Descrição atualizada do plano.",
                      "StartDate": "2024-01-15T00:00:00Z",
                      "EndDate": "2024-02-15T00:00:00Z"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Erro de validação de entrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "invalidIdFormatError": {
                    "summary": "Erro quando o ID fornecido não está no formato UUID válido.",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Formato de ID inválido.",
                      "status": 400,
                      "detail": "O formato do ID é inválido. Deve ser um UUID válido.",
                      "instance": "/api/Plan/Update/{Id}"
                    }
                  },
                  "invalidPlanData": {
                    "summary": "Erro quando os dados fornecidos para o plano são inválidos.",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Dados do plano inválidos.",
                      "status": 400,
                      "detail": "Ocorreu um erro ao validar os dados do plano. Verifique os detalhes.",
                      "instance": "/api/Plan/Update/{Id}",
                      "errors": {
                        "Name": [
                          "O nome do plano é obrigatório."
                        ],
                        "StartDate": [
                          "A data de início deve ser anterior à data de término."
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plano de treino não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "planNotFoundError": {
                    "summary": "Erro quando o plano de treino com o ID fornecido não é encontrado.",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Plano não encontrado.",
                      "status": 404,
                      "detail": "Plano de treino não encontrado com o ID fornecido.",
                      "instance": "/api/Plan/Update/{Id}"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Plan/Delete/{Id}": {
      "delete": {
        "tags": [
          "Plan"
        ],
        "summary": "Exclui um plano de treino específico.",
        "description": "Este endpoint permite a exclusão de um plano de treino específico utilizando o ID fornecido.",
        "parameters": [
          {
            "name": "Id",
            "in": "path",
            "required": true,
            "description": "O UUID do plano de treino que deseja excluir.",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "No Content - Plano de treino excluído com sucesso."
          },
          "400": {
            "description": "Erro de validação de entrada.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "invalidIdFormatError": {
                    "summary": "Erro quando o ID fornecido não está no formato UUID válido",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Formato de ID inválido.",
                      "status": 400,
                      "detail": "O formato do ID é inválido. Deve ser um UUID válido.",
                      "instance": "/api/Plan/Delete/{Id}"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Plano de treino não encontrado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetail"
                },
                "examples": {
                  "planNotFoundError": {
                    "summary": "Erro quando o plano de treino com o ID fornecido não é encontrado",
                    "value": {
                      "type": "https://tools.ietf.org/html/rfc9110#section-15.5.1",
                      "title": "Plano não encontrado.",
                      "status": 404,
                      "detail": "Plano de treino não encontrado com o ID fornecido.",
                      "instance": "/api/Plan/Delete/{Id}"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreatePlanCommand": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string",
            "description": "Nome do Plano de Treino.",
            "maxLength": 50,
            "minLength": 3,
            "nullable": true
          },
          "Description": {
            "type": "string",
            "description": "Descrição do Plano de Treino.",
            "maxLength": 240,
            "nullable": true
          },
          "StartDate": {
            "type": "string",
            "format": "date",
            "description": "Data de início do Plano de Treino.",
            "example": "2024-03-08"
          },
          "EndDate": {
            "type": "string",
            "format": "date",
            "description": "Data de término do Plano de Treino.",
            "example": "2024-04-08"
          },
          "Workouts": {
            "type": "array",
            "description": "Lista de IDs de Treinos associados ao Plano.",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "example": [
              "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "3fa85f64-5717-4562-b3fc-2c963f66afa7"
            ]
          }
        },
        "required": [
          "Name",
          "StartDate",
          "EndDate",
          "Workouts"
        ]
      },
...
      "UpdatePlanCommand": {
        "type": "object",
        "properties": {
          "Id": {
            "type": "string",
            "format": "uuid",
            "readOnly": true,
            "description": "Identificador único do Plano de Treino. Este campo é ignorado na requisição e definido internamente.",
            "example": "3fa85f64-5717-4562-b3fc-2c963f66afa6"
          },
          "Name": {
            "type": "string",
            "description": "Nome do Plano de Treino.",
            "maxLength": 50,
            "minLength": 3,
            "nullable": true
          },
          "Description": {
            "type": "string",
            "description": "Descrição do Plano de Treino.",
            "maxLength": 240,
            "nullable": true
          },
          "StartDate": {
            "type": "string",
            "format": "date",
            "description": "Data de início do Plano de Treino.",
            "example": "2024-03-08"
          },
          "EndDate": {
            "type": "string",
            "format": "date",
            "description": "Data de término do Plano de Treino.",
            "example": "2024-04-08"
          },
          "Workouts": {
            "type": "array",
            "description": "Lista de IDs de Treinos associados ao Plano.",
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "example": [
              "3fa85f64-5717-4562-b3fc-2c963f66afa6",
              "3fa85f64-5717-4562-b3fc-2c963f66afa7"
            ]
          }
        },
        "required": [
          "Name",
          "StartDate",
          "EndDate",
          "Workouts"
        ]
      },
...
      "PlanInfoViewModel": {
        "type": "object",
        "properties": {
          "Name": {
            "type": "string",
            "description": "Nome do Plano de Treino.",
            "example": "Plano de Força A"
          },
          "Description": {
            "type": "string",
            "description": "Descrição do Plano de Treino.",
            "example": "Foco em hipertrofia com treinos intensos."
          },
          "StartDate": {
            "type": "string",
            "format": "date-time",
            "description": "Data de início do Plano de Treino.",
            "example": "2024-01-01T00:00:00Z"
          },
          "EndDate": {
            "type": "string",
            "format": "date-time",
            "description": "Data de término do Plano de Treino.",
            "example": "2024-02-28T23:59:59Z"
          }
        }
      },
      "ProblemDetail": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "Um URI de referência sobre o tipo de problema."
          },
          "title": {
            "type": "string",
            "description": "Um título curto que descreve o problema."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "O código de status HTTP correspondente ao problema."
          },
          "detail": {
            "type": "string",
            "description": "Uma explicação detalhada do problema, geralmente incluindo informações da exceção."
          },
          "instance": {
            "type": "string",
            "description": "Um URI que identifica a instância específica do problema."
          },
          "errorCode": {
            "type": "string",
            "description": "Um código de erro específico que indica o tipo de falha da aplicação."
          },
          "traceId": {
            "type": "string",
            "description": "Um identificador único para rastrear a requisição específica no sistema."
          },
          "data": {
            "type": "object",
            "description": "Dados adicionais fornecidos pela exceção, geralmente valores chave do dicionário de dados da exceção."
          }
        },
        "required": [
          "title",
          "status"
        ],
        "example": {
          "type": "https://example.com/probs/out-of-credit",
          "title": "Erro interno no servidor.",
          "status": 500,
          "detail": "System.NullReferenceException = Object reference not set to an instance of an object.",
          "instance": "/minha-api/v1/resource",
          "errorCode": "500_ERR",
          "traceId": "00-ab123cd4ef567gh8-1234567890abcdef-01",
          "data": {
            "ChaveExemplo": "ValorExemplo"
          }
        }
      },
...
    }
  }
}